requirements.txt
flask
flask-cors
numpy
qutip
scipy
matplotlib
numba
pywebview
